Address some of last TODO
