+++ Ryujinx/Ui/MainWindow.cs
@@ -11,8 +11,6 @@ using System.Text;
 using System.Threading;
 using Ryujinx.Configuration;
 using System.Diagnostics;
-using OpenTK.Input;
-using Ryujinx.Ui.Input;
 using System.Threading.Tasks;
 using Utf8Json;
 using JsonPrettyPrinterPlus;
@@ -21,9 +19,6 @@ using Ryujinx.HLE.FileSystem;
 
 
 using GUI = Gtk.Builder.ObjectAttribute;
-using Ryujinx.HLE.Input;
-using Ryujinx.Configuration.Hid;
-using Ryujinx.HLE.HOS.SystemState;
 
 namespace Ryujinx.Ui
 {
@@ -206,7 +201,8 @@ namespace Ryujinx.Ui
         {
             HLE.Switch instance = new HLE.Switch(_renderer, _audioOut);
 
-            instance.Initialize((SystemLanguage)ConfigurationState.Instance.System.Language.Value, ConfigurationState.Instance.System.EnableMulticoreScheduling, ConfigurationState.Instance.System.EnableDockedMode, ConfigurationState.Instance.Graphics.EnableVsync, ConfigurationState.Instance.System.EnableFsIntegrityChecks, ConfigurationState.Instance.System.FsGlobalAccessLogMode, ConfigurationState.Instance.System.IgnoreMissingServices);
+            instance.Initialize();
+
             return instance;
         }
 
@@ -340,35 +336,9 @@ namespace Ryujinx.Ui
             }
         }
 
-        // TODO: Make HLE.Switch handle the state itself + reloading
-        private static void ConfigureHid()
-        {
-            _device.Hid.InitializePrimaryController(ConvertControllerTypeToState(ConfigurationState.Instance.Hid.ControllerType));
-            _device.Hid.InitializeKeyboard();
-        }
-
-        private static ControllerStatus ConvertControllerTypeToState(ControllerType controllerType)
-        {
-            switch (controllerType)
-            {
-                case ControllerType.Handheld:
-                    return ControllerStatus.Handheld;
-                case ControllerType.NpadLeft:
-                    return ControllerStatus.NpadLeft;
-                case ControllerType.NpadRight:
-                    return ControllerStatus.NpadRight;
-                case ControllerType.NpadPair:
-                    return ControllerStatus.NpadPair;
-                case ControllerType.ProController:
-                    return ControllerStatus.ProController;
-                default:
-                    throw new NotImplementedException();
-            }
-        }
-
         private static void CreateGameWindow()
         {
-            ConfigureHid();
+            _device.Hid.InitializePrimaryController(ConfigurationState.Instance.Hid.ControllerType);
 
             using (_screen = new GlScreen(_device, _renderer))
             {
